move check for invalid characters to linked files editor
Fixes #12212

some cleanup

+++ src/main/java/org/jabref/gui/fieldeditors/LinkedFilesEditor.java
@@ -253,13 +253,10 @@ public class LinkedFilesEditor extends HBox implements FieldEditorFX {
         listView.addEventFilter(KeyEvent.KEY_PRESSED, event -> {
             Optional<KeyBinding> keyBinding = preferences.getKeyBindingRepository().mapToKeyBinding(event);
             if (keyBinding.isPresent()) {
-                switch (keyBinding.get()) {
-                    case DELETE_ENTRY:
-                        deleteAttachedFilesWithConfirmation();
-                        event.consume();
-                        break;
-                    default:
-                        // Pass other keys to children
+                // Pass other keys to children
+                if (keyBinding.get() == KeyBinding.DELETE_ENTRY) {
+                    deleteAttachedFilesWithConfirmation();
+                    event.consume();
                 }
             }
         });
@@ -287,7 +284,7 @@ public class LinkedFilesEditor extends HBox implements FieldEditorFX {
 
     @FXML
     private void addNewFile() {
-        dialogService.showCustomDialogAndWait(new LinkedFileEditDialog()).ifPresent(newLinkedFile -> {
+        dialogService.showCustomDialogAndWait(new LinkedFileEditDialog()).filter(file -> !file.isEmpty()).ifPresent(newLinkedFile -> {
             viewModel.addNewLinkedFile(newLinkedFile);
         });
     }
